优化rm bash

+++ scripts/develop.js
@@ -11,8 +11,11 @@ if (!shell.which('fis3')) {
     console.log(chalk.red('请先安装fis3'));
     return;
 }
-
-shell.rm('-rf', configBuild.output + configBuild.static_root + '/*');
-shell.rm('-rf', configBuild.output + configBuild.tpl_root+ '/*');
+if (shell.test('-d', configBuild.output + configBuild.static_root)) {
+    shell.rm('-rf', configBuild.output + configBuild.static_root + '/*');
+}
+if (shell.test('-d', configBuild.output + configBuild.tpl_root)) {
+    shell.rm('-rf', configBuild.output + configBuild.tpl_root+ '/*');
+}
 shell.exec('fis3 release -cd ' + configBuild.output);
